+++ tests/test_cli.py
@@ -2,14 +2,15 @@ import subprocess
 import sys
 
 from pymdgen.cli import run
+from helpers import write_expected
 
 
 def test_cli(expected_docs_md):
     output = subprocess.check_output(["pymdgen", "pymdgen.test_module"])
-    print(type(output))
     if isinstance(output, bytes):
         output = output.decode("unicode_escape")
-    assert output == expected_docs_md
+
+    assert output == expected_docs_md+"\n"
 
 
 def test_run(expected_docs_list):