+++ tests/test_gen_docs.py
@@ -21,16 +21,17 @@ class class_a(object):
 
 def test_doc_func(expected_docs_doc_func_md):
     output = doc_func("func_a", func_a)
-    # write_expected("doc_func.md", output)
+    #write_expected("doc_func.md", output)
     assert "\n".join(output) == expected_docs_doc_func_md
 
 
 def test_doc_class(expected_docs_doc_class_md):
     output = doc_class("class_a", class_a)
-    # write_expected("doc_class.md", output)
+    #write_expected("doc_class.md", output)
     assert "\n".join(output) == expected_docs_doc_class_md
 
 
 def test_doc_module(expected_docs_list):
     output = doc_module("pymdgen.test_module", section_level=3)
+    #write_expected("md", output)
     assert output == expected_docs_list